deal with type long problem

+++ push-swap/turk_sort/sort_algo/sort_algo.h
@@ -6,7 +6,7 @@
 /*   By: ytoshihi <ytoshihi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 09:13:12 by ytoshihi          #+#    #+#             */
-/*   Updated: 2024/04/02 14:13:40 by ytoshihi         ###   ########.fr       */
+/*   Updated: 2024/04/02 17:11:02 by ytoshihi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,6 +35,6 @@ int		push_two_num(t_list **a, t_list **b);
 void	push_with_sort(t_list **a, t_list **b);
 void	sort_three_nums(t_list **a);
 void	push_all_from_b(t_list **a, t_list **b);
-int		search_min_num(t_list *b, int num);
+long	search_min_num(t_list *b, long num);
 
 #endif
